Cambios

+++ Function.java
@@ -4,22 +4,22 @@ import java.util.LinkedList;
 public class Function {
   public String name;
   public LinkedList<Object> body;
-  public HashMap<String, Integer>  params;
+  public HashMap<String, Double>  params;
 
-  public Function(String nombre, LinkedList<Object> cuerpo, HashMap<String, Integer> parametros){
-      name = nombre;
-      body = cuerpo;
-      params = parametros;
+  public Function(String nombre, LinkedList<Object> cuerpo, HashMap<String, Double> parametros){
+    name = nombre;
+    body = cuerpo;
+    params = parametros;
 
   }
 
-  public void setparametro(String nombrepar, int par){
+  public void setparametro(String nombrepar, double par){
     params.remove(nombrepar);
     params.put(nombrepar, par);
-  }
+}
 
-  public Integer getparametro(String par){
-    Integer valor = params.get(par);
+  public Double getparametro(String par){
+    Double valor = params.get(par);
     return valor;
   }
